Update to SOwISC12to30E3r3
Switch SO12to60 to SO12to30 with a base resolution that is now
uniformly 30 km.
