+++ compass/ocean/suites/sowisc12to30.txt
@@ -1,0 +1,5 @@
+ocean/global_ocean/SOwISC12to30/mesh
+ocean/global_ocean/SOwISC12to30/WOA23/init
+ocean/global_ocean/SOwISC12to30/WOA23/performance_test
+ocean/global_ocean/SOwISC12to30/WOA23/dynamic_adjustment
+ocean/global_ocean/SOwISC12to30/WOA23/files_for_e3sm
