--- compass/ocean/suites/sowisc12to60.txt
@@ -1,5 +1,0 @@
-ocean/global_ocean/SOwISC12to60/mesh
-ocean/global_ocean/SOwISC12to60/WOA23/init
-ocean/global_ocean/SOwISC12to60/WOA23/performance_test
-ocean/global_ocean/SOwISC12to60/WOA23/dynamic_adjustment
-ocean/global_ocean/SOwISC12to60/WOA23/files_for_e3sm
