+++ compass/ocean/tests/global_ocean/mesh/__init__.py
@@ -16,7 +16,7 @@ from compass.ocean.tests.global_ocean.mesh.qu import (
     QUMeshFromConfigStep,
 )
 from compass.ocean.tests.global_ocean.mesh.rrs6to18 import RRS6to18BaseMesh
-from compass.ocean.tests.global_ocean.mesh.so12to60 import SO12to60BaseMesh
+from compass.ocean.tests.global_ocean.mesh.so12to30 import SO12to30BaseMesh
 from compass.ocean.tests.global_ocean.mesh.wc14 import WC14BaseMesh
 from compass.ocean.tests.global_ocean.metadata import (
     get_author_and_email_from_git,
@@ -98,8 +98,8 @@ class Mesh(TestCase):
             base_mesh_step = ARRM10to60BaseMesh(self, name=name, subdir=subdir)
         elif mesh_name in ['RRS6to18', 'RRSwISC6to18']:
             base_mesh_step = RRS6to18BaseMesh(self, name=name, subdir=subdir)
-        elif mesh_name in ['SO12to60', 'SOwISC12to60']:
-            base_mesh_step = SO12to60BaseMesh(self, name=name, subdir=subdir)
+        elif mesh_name in ['SO12to30', 'SOwISC12to30']:
+            base_mesh_step = SO12to30BaseMesh(self, name=name, subdir=subdir)
         elif mesh_name in ['FRIS01to60', 'FRISwISC01to60']:
             base_mesh_step = FRIS01to60BaseMesh(self, name=name, subdir=subdir)
         elif mesh_name in ['FRIS02to60', 'FRISwISC02to60']:
