+++ compass/ocean/tests/global_ocean/mesh/so12to30/dynamic_adjustment.yaml
