+++ compass/ocean/tests/global_ocean/mesh/so12to30/high_res_region.geojson
@@ -4,7 +4,7 @@
     {
       "type": "Feature",
       "properties": {
-        "name": "SO12to60 high res region",
+        "name": "SO12to30 high res region",
         "component": "ocean",
         "object": "region",
         "author": "Xylar Asay-Davis"
