+++ compass/ocean/tests/global_ocean/mesh/so12to30/namelist.split_explicit_ab2
