+++ compass/ocean/tests/global_ocean/mesh/so12to30/so12to30.cfg
@@ -33,23 +33,20 @@ prefix = SO
 # a description of the mesh and initial condition
 mesh_description = MPAS Southern Ocean regionally refined mesh for E3SM version
                    ${e3sm_version} with enhanced resolution (${min_res} km) around
-                   Antarctica, 45-km resolution in the mid southern latitudes,
-                   30-km resolution in a 15-degree band around the equator, 60-km
-                   resolution in northern mid latitudes, 30 km in the north
-                   Atlantic and 35 km in the Arctic.  This mesh has <<<levels>>>
+                   Antarctica and ${max_res} km elsewhere.  This mesh has <<<levels>>>
                    vertical levels and includes cavities under the ice shelves
                    around Antarctica.
 # E3SM version that the mesh is intended for
 e3sm_version = 3
 # The revision number of the mesh, which should be incremented each time the
 # mesh is revised
-mesh_revision = 1
+mesh_revision = 3
 # the minimum (finest) resolution in the mesh
 min_res = 12
 # the maximum (coarsest) resolution in the mesh, can be the same as min_res
-max_res = 60
+max_res = 30
 # The URL of the pull request documenting the creation of the mesh
-pull_request = https://github.com/MPAS-Dev/compass/pull/669
+pull_request = https://github.com/MPAS-Dev/compass/pull/807
 
 
 # config options related to initial condition and diagnostics support files
