--- compass/ocean/tests/global_ocean/mesh/so12to60/atlantic.geojson
@@ -1,97 +1,0 @@
-{
-  "type": "FeatureCollection",
-  "features": [
-    {
-      "type": "Feature",
-      "properties": {
-        "name": "Atlantic region",
-        "component": "ocean",
-        "object": "region",
-        "author": "Xylar Asay-Davis"
-      },
-      "geometry": {
-        "type": "Polygon",
-        "coordinates": [
-          [
-            [
-              -97.3828125,
-              85.05112877979998
-            ],
-            [
-              -102.3046875,
-              40.17887331434696
-            ],
-            [
-              -102.3046875,
-              23.241346102386135
-            ],
-            [
-              -93.1640625,
-              15.623036831528264
-            ],
-            [
-              -85.78125,
-              13.581920900545844
-            ],
-            [
-              -83.583984375,
-              9.535748998133627
-            ],
-            [
-              -81.2109375,
-              8.059229627200192
-            ],
-            [
-              -79.013671875,
-              9.795677582829743
-            ],
-            [
-              -75.9375,
-              5.61598581915534
-            ],
-            [
-              -77.6953125,
-              0
-            ],
-            [
-              16.171875,
-              0
-            ],
-            [
-              27.773437499999996,
-              26.745610382199022
-            ],
-            [
-              37.96875,
-              32.24997445586331
-            ],
-            [
-              39.7265625,
-              39.36827914916014
-            ],
-            [
-              32.6953125,
-              53.9560855309879
-            ],
-            [
-              37.6171875,
-              61.438767493682825
-            ],
-            [
-              25.664062500000004,
-              68.26938680456564
-            ],
-            [
-              24.609375,
-              85.05112877979998
-            ],
-            [
-              -97.3828125,
-              85.05112877979998
-            ]
-          ]
-        ]
-      }
-    }
-  ]
-}
